+++ scene/src/components/button-chip/index.ts
@@ -1,0 +1,1 @@
+export { default as ButtonChip } from './ButtonChip';
