+++ scene/src/components/button-icon/index.ts
@@ -1,0 +1,1 @@
+export { default as ButtonIcon } from './ButtonIcon';
