+++ scene/src/components/button-text-icon/index.ts
@@ -1,0 +1,1 @@
+export { default as ButtonTextIcon } from './ButtonTextIcon';
