+++ scene/src/components/button-text/ButtonText.tsx
@@ -4,10 +4,8 @@ import ReactEcs, {
   UiEntity,
   type UiTransformProps
 } from '@dcl/sdk/react-ecs'
-import { type Icon } from '../utils/definitions'
-import { getBackgroundFromAtlas } from '../utils/ui-utils'
 
-function TextIconButton(props: {
+function ButtonText(props: {
   // Events
   onMouseEnter?: Callback
   onMouseLeave?: Callback
@@ -18,25 +16,26 @@ function TextIconButton(props: {
   // Text
   value: string
   fontSize: number
-  icon: Icon
-  iconSize?: number
   fontColor?: Color4
-  iconColor?: Color4
-  direction?: 'row' | 'column'
+  // Status
+  isLoading?: boolean
 }): ReactEcs.JSX.Element | null {
   //   const ICON_MARGIN = Math.max(canvasInfo.height * 0.01, 2)
   return (
     <UiEntity
       uiTransform={{
         // padding: props.fontSize * 0.3,
-        // margin: { bottom: props.fontSize * 0.3, top: props.fontSize * 0.3 },
-        flexDirection: props.direction ?? 'row',
+        // margin: props.fontSize * 0.3,
+        flexDirection: 'row',
         justifyContent: 'center',
         alignItems: 'center',
+        width: 'auto',
+        height: 'auto',
         ...props.uiTransform
       }}
       uiBackground={{
         color: props.backgroundColor ?? { ...Color4.White(), a: 0 },
+
         textureMode: 'nine-slices',
         texture: {
           src: 'assets/images/backgrounds/rounded.png'
@@ -52,18 +51,6 @@ function TextIconButton(props: {
       onMouseEnter={props.onMouseEnter}
       onMouseLeave={props.onMouseLeave}
     >
-      {/* ICON */}
-
-      <UiEntity
-        uiTransform={{
-          width: props.iconSize ?? 2 * props.fontSize,
-          height: props.iconSize ?? 2 * props.fontSize
-        }}
-        uiBackground={{
-          ...getBackgroundFromAtlas(props.icon),
-          color: props.iconColor
-        }}
-      />
       {/* TEXT */}
       <UiEntity
         uiTransform={{
@@ -75,9 +62,25 @@ function TextIconButton(props: {
           fontSize: props.fontSize,
           color: props.fontColor ?? Color4.White()
         }}
-      ></UiEntity>
+      >
+        {/* ICON */}
+        {props.isLoading === true && (
+          <UiEntity
+            uiTransform={{
+              width: props.fontSize,
+              height: props.fontSize,
+              positionType: 'absolute',
+              position: { top: '25%', left: -1.25 * props.fontSize }
+            }}
+            uiBackground={{
+              textureMode: 'stretch',
+              texture: { src: 'assets/images/Spinner.png' }
+            }}
+          />
+        )}
+      </UiEntity>
     </UiEntity>
   )
 }
 
-export default TextIconButton
+export default ButtonText
