+++ scene/src/components/button-text/index.ts
@@ -1,0 +1,1 @@
+export { default as ButtonText } from './ButtonText';
