+++ scene/src/components/chat-message/index.ts
@@ -1,0 +1,1 @@
+export { default as ChatMessage } from './ChatMessage';
