+++ scene/src/components/dropdown-field/index.ts
@@ -1,0 +1,1 @@
+export { default as DropdownField } from './DropdownField';
