+++ scene/src/components/dropdown-styled/index.ts
@@ -1,0 +1,1 @@
+export { default as StyledDropdown } from './StyledDropdown';
