+++ scene/src/components/friend-invitation/index.ts
@@ -1,0 +1,1 @@
+export { default as FriendInvitation } from './FriendInvitation';
