+++ scene/src/components/friend-item/index.ts
@@ -1,0 +1,1 @@
+export { default as FriendItem } from './FriendItem';
