+++ scene/src/components/input-field/index.ts
@@ -1,0 +1,1 @@
+export { default as InputField } from './InputField';
