+++ scene/src/components/invitation-item/index.ts
@@ -1,0 +1,1 @@
+export { default as InvitationItem } from './InvitationItem';
