+++ scene/src/components/slider/index.ts
@@ -1,0 +1,1 @@
+export { default as Slider } from './Slider';
