+++ scene/src/ui/main-hud/actionPopUp.tsx
@@ -1,8 +1,8 @@
 import { UiCanvasInformation, engine } from '@dcl/sdk/ecs'
 import { Color4 } from '@dcl/sdk/math'
 import ReactEcs, { type Callback, Label, UiEntity } from '@dcl/sdk/react-ecs'
-import IconButton from '../../components/button-icon/iconButton'
-import TextButton from '../../components/button-text/textButton'
+import IconButton from '../../components/button-icon/ButtonIcon'
+import TextButton from '../../components/button-text/ButtonText'
 import { type UIController } from '../../controllers/ui.controller'
 import {
   ALMOST_BLACK,
