+++ scene/src/ui/profile/index.tsx
@@ -4,9 +4,9 @@ import { Color4 } from '@dcl/sdk/math'
 import ReactEcs, { UiEntity } from '@dcl/sdk/react-ecs'
 import ChipButton from '../../components/chipButton'
 import DropdownField from '../../components/dropdownField'
-import IconButton from '../../components/button-icon/iconButton'
+import IconButton from '../../components/button-icon/ButtonIcon'
 import InputField from '../../components/inputField'
-import TextButton from '../../components/button-text/textButton'
+import TextButton from '../../components/button-text/ButtonText'
 import TextIconButton from '../../components/textIconButton'
 import { type UIController } from '../../controllers/ui.controller'
 import {
